feat/#134 테스트 완료 이미지 컨버터 적용
