+++ src/views/profile/ui/EditProfileViewTest.tsx
@@ -6,7 +6,7 @@ import { ChevronLeft } from 'lucide-react';
 import { toast } from 'sonner';
 import { Button } from '@/shared/ui/Button';
 import { FallbackScreen } from '@/widgets/common/FallbackScreen';
-import { ProfileImageUploader } from '@/features/edit-profile/ui/ProfileImageUploaderTest';
+import { ProfileImageUploader } from '@/features/edit-profile/ui/ProfileImageUploader';
 import { ProfileInfoEditor } from '@/features/edit-profile/ui/ProfileInfoEditorTest';
 import { uploadImagesList } from '@/features/update-user/api/uploadImagesList';
 import { updateProfile, UpdateProfileRequest } from '@/features/edit-profile/model/updateProfile';
